requirements.txt
# Core dependencies
python-dotenv>=0.19.0
openai>=0.27.0

# PDF processing
PyPDF2>=2.0.0
pdfminer.six>=20221105

# Vector database options
chromadb>=0.4.13
pydantic<2.0.0

# Embedding libraries (one of these will be used)
llama-index>=0.8.11
langchain>=0.0.267

# Knowledge graph
networkx>=2.8.8
pyvis>=0.3.2

# Utility
numpy>=1.24.3
matplotlib>=3.7.1